Create VM based on image
In order to create a VM based on image, the user should create an image
under Compute Resource --> Create Image, and provide the name of the
image on the registry.

The image will be set as the registryDisk of the vm (should be renamed
to containerDisk in a following PR).

One property of 'registryDisk' is missing ('path') and should be added
in a following PR as well